Add files via upload
Changed the primary color from blue to teal in style.css. Mainly buttons and alerts were affected. This involved lots of find/replace with slightly different shades of blue and replacing them with slightly different shades of teal. Hopefully it will be easier to create new widgets on the website because the colors will be correct and consistent.